+++ mplang/core/interp.py
@@ -26,7 +26,8 @@ from abc import abstractmethod
 from collections.abc import Sequence
 from typing import Any, cast
 
-from mplang.core.base import MPContext, MPObject, MPType, TensorLike
+from mplang.core.mpobject import MPContext, MPObject
+from mplang.core.mptype import MPType, TensorLike
 from mplang.core.trace import TracedFunction
 from mplang.expr.ast import Expr, VariableExpr
 from mplang.utils.func_utils import var_demorph, var_morph
