+++ mplang/frontend/builtin.py
@@ -16,7 +16,8 @@ from typing import Any
 
 from jax.tree_util import PyTreeDef, tree_flatten
 
-from mplang.core.base import MPObject, TensorInfo
+from mplang.core.mpobject import MPObject
+from mplang.core.mptype import TensorInfo
 from mplang.core.pfunc import PFunction
