+++ mplang/frontend/jax_cc.py
@@ -21,7 +21,8 @@ import jax
 import jax.numpy as jnp
 from jax.tree_util import PyTreeDef, tree_flatten
 
-from mplang.core.base import MPObject, TensorInfo
+from mplang.core.mpobject import MPObject
+from mplang.core.mptype import TensorInfo
 from mplang.core.pfunc import PFunction, get_fn_name
 from mplang.utils.func_utils import normalize_fn
