+++ mplang/runtime/simulation.py
@@ -26,9 +26,11 @@ import spu.libspu as libspu
 from mplang.backend.builtin import BuiltinHandler
 from mplang.backend.spu import SpuHandler
 from mplang.backend.stablehlo import StablehloHandler
-from mplang.core.base import Mask, MPObject, MPType, TensorLike
 from mplang.core.interp import InterpContext, InterpVar
+from mplang.core.mask import Mask
 from mplang.core.mpir import Reader, Writer
+from mplang.core.mpobject import MPObject
+from mplang.core.mptype import MPType, TensorLike
 from mplang.expr.ast import Expr
 from mplang.expr.evaluator import Evaluator
 from mplang.runtime.grpc_comm import LinkCommunicator
