+++ tests/backend/test_builtin.py
@@ -19,7 +19,7 @@ import numpy as np
 import pytest
 
 from mplang.backend.builtin import BuiltinHandler
-from mplang.core.base import TensorInfo
+from mplang.core.mptype import TensorInfo
 from mplang.core.pfunc import PFunction
