+++ tests/core/test_mptype.py
@@ -15,9 +15,9 @@
 import numpy as np
 import pytest
 
-from mplang.core.base import MPKind, MPType
 from mplang.core.dtype import DATE, FLOAT32, FLOAT64, INT32, INT64, JSON, STRING
 from mplang.core.mask import Mask
+from mplang.core.mptype import MPKind, MPType
 from mplang.core.relation import RelationSchema
