+++ tests/expr/test_visitor.py
@@ -14,8 +14,9 @@
 
 import pytest
 
-from mplang.core.base import Mask, TensorInfo
 from mplang.core.dtype import FLOAT32, INT32
+from mplang.core.mask import Mask
+from mplang.core.mptype import TensorInfo
 from mplang.expr import ConstExpr, Expr, ExprTransformer, Printer
 from mplang.expr.ast import RankExpr
