+++ tests/frontend/test_spu.py
@@ -17,7 +17,7 @@ import numpy as np
 import pytest
 import spu.libspu as libspu
 
-from mplang.core.base import TensorInfo
+from mplang.core.mptype import TensorInfo
 from mplang.frontend.spu import SpuFE, Visibility
